# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: WatchMe API
  description: API para o sistema de catálogo de filmes WatchMe
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.watchme.com
    description: Production server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Movie:
      type: object
      properties:
        id:
          type: integer
          example: 550
        title:
          type: string
          example: "Fight Club"
        overview:
          type: string
          example: "Um homem deprimido que sofre de insônia..."
        poster_path:
          type: string
          example: "https://image.tmdb.org/t/p/w500/poster.jpg"
        backdrop_path:
          type: string
          example: "https://image.tmdb.org/t/p/w500/backdrop.jpg"
        release_date:
          type: string
          format: date
          example: "1999-10-15"
        vote_average:
          type: number
          format: float
          example: 8.4

    MovieDetails:
      allOf:
        - $ref: '#/components/schemas/Movie'
        - type: object
          properties:
            genres:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
            runtime:
              type: integer
            status:
              type: string
            tagline:
              type: string

    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string

paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login do usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login bem sucedido
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /movies/popular:
    get:
      tags:
        - Movies
      summary: Lista filmes populares
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de filmes retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /movies/{id}:
    get:
      tags:
        - Movies
      summary: Obtém detalhes de um filme
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID do filme no TMDb
      responses:
        '200':
          description: Detalhes do filme retornados com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetails'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Filme não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error' 